requirements.txt
asgiref==3.2.3
Django==3.0.3
django-extensions==2.2.8
django-phone-field==1.8.0
djangorestframework==3.11.0
pytz==2019.3
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
